+++ lib/restful_json/controller/nil_param_values.rb
@@ -9,9 +9,9 @@ module RestfulJson
       NILS = ['NULL'.freeze, 'null'.freeze, 'nil'.freeze].to_set
 
       # Converts request param values 'NULL', 'null', and 'nil' to nil.
-      def convert_request_param_value_for_filtering(attr_sym, value)
-        value = super || value
-        value && NILS.include?(value) ? nil : value
+      def convert_request_param_value_for_filtering(attr_sym, param_value)
+        param_value = super || param_value
+        param_value && NILS.include?(param_value) ? nil : param_value
       end
 
     end