+++ lib/restful_json/controller/validation_errors.rb
@@ -1,5 +1,5 @@
-# Before every controller action, calls authorize! with the action and model class
-# for easy integration with authorizers like CanCan, etc.
+# Instead of having to handle @my_model_name.errors in the view, this returns validation errors
+# via rendering { errors: record.errors } for all formats except html.
 module RestfulJson
   module Controller
     module ValidationErrors
@@ -7,20 +7,20 @@ module RestfulJson
 
       include ::RestfulJson::Controller
 
-      def render_create_invalid(value)
-        render_validation_errors(value)
+      def render_create_invalid(record)
+        render_validation_errors(record)
       end
 
-      def render_update_invalid(value)
-        render_validation_errors(value)
+      def render_update_invalid(record)
+        render_validation_errors(record)
       end
 
-      def render_validation_errors(value)
+      def render_validation_errors(record)
         content_type = request.formats.first.to_s.reverse.split('/')[0].split('-')[0].reverse
         # use implicit rendering for html
-        return value if request.format.html?
+        return record if request.format.html?
         respond_to do |format|
-          format.any { render content_type.to_sym => { errors: value.errors }, status: 422 }
+          format.any { render content_type.to_sym => { errors: record.errors }, status: 422 }
         end
       end