+++ spec/dummy/app/controllers/barfoos_controller.rb
@@ -3,7 +3,6 @@ class BarfoosController < ApplicationController
   include RestfulJson::Controller::StatusAndLocation
   respond_to :json
   
-  query_for some_action: ->(t,q) {q.where(:status => 2)}
+  query_for some_action: ->(q) {q.where(:status => 2)}
   including :foo
 end
-